Process cache identities from ViewModels
When ViewModels instantiated by the view model registry implement the
IndentityInterface, the cache tags will be added to the HTTP response
X-Magento-Tags header that is used to set the cache tags for full page
cache records.

This is done for both regular pages as well as Varnish ESI request responses.

This change allows to use regular Template blocks instead of more
specific block classes, that often are more expensive to instantiate.