Add only return identities if the current product is loaded
The view model is also used on the product list page, where instead of
calling getProduct(), the product is passed as a method argument.
In these cases the view model should not return any identities, as they
are already taken care of by the product block that is used to render
the page.

This commit also adds a method to has an array of scalar values to be
used as a block cache key.