Pass product instance as argument
Relying on the product to be set on the view model, or on the current
product registry to be populated, introduces temporal coupling.

This lead to issues after the 1.1.2 release where customizations to the
pdp that changed the order or product templates being rendered.
(For example Error: Call to a member function getPriceInfo() on null in
vendor/hyva-themes/magento2-theme-module/src/ViewModel/ProductPrice.php:84)

To preserve backward compatibility the product argument is optional.
The public setProduct method and the protected getProduct method have
been marked as deprecated, but still are kept for backward
compatibility.

All use cases in the magento2-default-theme have been updated
accordingly.