Add useAnchorAttribute property to load category
Adds the useAnchorAttribute property to give an option and load the
whole category instead of creating a category with only an id. This
enables to load the products of the category based on whether it is an
anchor category or not. With the property being false the category will
always be treated as not an anchor due to the implementation of
\Magento\Catalog\Model\ResourceModel\Product\Collection::addCategoryFilter.